requirements.txt
PyYAML==6.0

argparse==1.4.0
beautifulsoup4==4.11.1
ffmpeg==1.4
moviepy==1.0.3
regex==2022.7.25
requests==2.28.1
torch==1.12.1
yt_dlp==2022.8.8

git+https://github.com/openai/whisper.git